Enhance DTOs and repositories for new specification fields
Added new properties and mappings in DTOs for CoreInvoiceModel and ExtensionComponentModelElement. Introduced an anonymous access endpoint in SpecificationsController. Updated SpecificationCoreDto and SpecificationExtensionComponentDto with new fields and constructors. Enhanced mapping in SpecificationProfile. Modified repositories for eager loading and ordering based on new properties. Overall improvements for better handling of specifications and related components.